Wrote ThreadController::write()
When the user submits to thread/write, write() will be called. The
thread id is obtained, and the hidden input, page_next is checked
for its value. If its value is write_end, a new comment instance
will be created and written to the database. Otherwise, an
exception is thrown. The view is then rendered.