Added validate_name() helper function
validate_name() checks if the given string contains only letters
and spaces. Returns a true value if it is and a false value other-
wise. Used preg_match() because the ctype_alpha() checks for
letters, but not spaces (/[A-Za-z]/).

Comment model validation rules have been updated as well to use the
validate_name().

- If there is another function that could do the same thing,
  I would replace preg_match().