Added trim_collapse() helper function
trim_collapse() removes the leading and trailing whitespace from a
string and  collapses adjacent whitespaces into one. It returns
the resulting string.

Whitespace is preserved when the name or thread titles are stored
in the database. I thought it looked ugly and it might be a problem
during searching, so I did this.

- trim_collapse() is not used on the comment body because I think
  should have more freedom in formatting his comments