Added case to show registration page
The value of the hidden input, page_next determines if the user
should be shown the registration form, or if the user is attempting
to create a new user. If page_next doesn't have a value, it means
that the user haven't filled a form yet, and must be shown to the
registration page.

- TODO: create a NotFoundException class.