Changed default values of registration form
The default values of the fields in the registration form are
defined by the properties of the User instance. However, if
the user haven't filled up a form, these properties will be
undefined. Because of this, error messages become the default
values in the input fields.

By using Param::get(), the input that the user submitted can
be retrieved, and the default values can also remain blank if
the user haven't filled up the form yet.