User::authenticate() no longer throws RecordNotFoundException
The way it was, if the user gives a wrong username, the method
would thrown an exception. If the user gives a valid username,
but an incorrect password, the method would return false.
I think giving a wrong username or password should result to the
same outcome. Now, the method would return false if either the
username or password is wrong.