Fixed: 1st page is clickable if currently on page 1
If the url does not contain the page query, the printPageLinks
the first page link is always clickable even if you're currently
on the first page. This is because printPageLinks() rely on the
value of 'page' query. By giving a default value if it's not set,
the problem is solved.