Updated thread creation to store category id
Saved category id to thread table

Thread::create() throws CategoryException

When the category input does not match any category, a CategoryException
will be thrown. The constant, Thread::ERR_CATEGORY is actually the
code for a foreign key constraint failure. However, in the Thread
model, that kind of failure can only exist from the category_id column,
so I named it as such.