Fixed ranking
Order of items are not correct.
Now, Comment::countToday() will return an array of the count of
today's comments in each thread. Thread::getTrending() will return
an array of threads based on the array returned by Comment::countToday().
It will also sort the threads based on the number of new comments
and creation date.